backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt,argon2]>=1.7.4
bcrypt>=4.0.1
argon2-cffi>=23.1.0


# AI / LLM
groq>=0.11.0
openai>=1.50.0
httpx>=0.27.0
huggingface-hub>=0.25.0

# RAG Pipeline
faiss-cpu>=1.7.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
tiktoken>=0.7.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.3
requests>=2.32.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.0.0
python-dotenv>=1.0.1
aiofiles>=24.1.0

